Add a utility library and number of utility programs
Add a utility library to do parsing of cellserv database and DNS lookups
for server details.  This will be of use to the kafs-utils package as well.
Parts of it may also be generally useful, so it may get moved into
keyutils.

Add a number of utility programs:

 (1) kafs-check-config: Sysadmin tool.  Check the cellserv database and
     allow it to be dumped.

 (2) kafs-preload: Systemd auxiliary program.  Preload the in-kernel cell
     database from the configuration.

 (3) dns_afsdb: Request-key upcall program.  Fetch the VL server list and
     addresses lists for a cell on behalf of the kernel.

Add a systemd service to configure the kafs dynamic root by preloading cell
names into the cell database.

Signed-off-by: David Howells <dhowells@redhat.com>