Cleanup systemd install
This patch upgrades the dh compatibility to 11 and hence uses
dh_installsystemd.  Makes systemd installing slightly more
straight-forward.

Notably, this patch also stops changing the name of the kafs-config.service
to kafs-client.service--the config name seems more appropriate for the
action and sticks with the naming of the install done by the Makefile.